% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonds.R
\name{findBonds}
\alias{findBonds}
\title{Detects specific functional groups in a molecule.}
\usage{
findBonds(smiles, targetGroup)
}
\arguments{
\item{smiles}{A SMILES-formatted molecule string.}

\item{targetGroup}{A predefined character vector of functional groups.}
}
\value{
A tibble containing bond block and atom block information for each functional group detected.
}
\description{
Detects specific functional groups in a molecule (SMILES format) and returns a table containing detailed SDF information. Slower than smartBonds(), but provides more detailed information.
}
