% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPCFrags.R
\name{fragSites}
\alias{fragSites}
\title{Quickly detects specific functional groups in a molecule}
\usage{
fragSites(smiles, grps)
}
\arguments{
\item{smiles}{A SMILES-formatted molecule string.}

\item{grps}{A predefined character vector of functional groups.}
}
\value{
A named character vector.
}
\description{
Quickly detects specific functional groups in a molecule
}
