% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSDF.R
\name{makeSDF}
\alias{makeSDF}
\title{Construct a new SDFset object for a single molecule from modified atom block and bond block tables (??"SDFset-class" for details).}
\usage{
makeSDF(prodTables, sdf)
}
\arguments{
\item{prodTables}{A list of two matrices: an atoms table and a bonds table}

\item{sdf}{The unmodified SDF object from which the original atom block and bond block tables were obtained}
}
\value{
A new SDFset object containing one molecule.
}
\description{
Construct a new SDFset object for a single molecule from modified atom block and bond block tables (??"SDFset-class" for details).
}
