% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPCFrags.R
\name{OPCproducts.Frag}
\alias{OPCproducts.Frag}
\title{Predict common fragments of OPCs (organophosphorus compounds).}
\usage{
OPCproducts.Frag(smilesList)
}
\arguments{
\item{smilesList}{Input character vector of organophosphorus compounds. Must be in SMILES (Simplified Molecular Input Line Entry Specification) format.}
}
\value{
Tibble containing input SMILES and predicted positive/negative fragment m/z values for each.
}
\description{
Predict common fragments of OPCs (organophosphorus compounds).
}
