% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPCFrags.R
\name{predict.Frag}
\alias{predict.Frag}
\title{Predict fragments from specified reaction events on a single molecule.}
\usage{
\method{predict}{Frag}(cmpd, grp, rxnList, rxnSave)
}
\arguments{
\item{cmpd}{SMILES string representing a single molecule.}

\item{grp}{Single-element named character vector containing a functional group. Name represents atom indices. Output of smartBonds().}

\item{rxnList}{Character vector of allowed reaction types.}

\item{rxnSave}{Character vector containing saved reactions preceding this one in sequence.}
}
\value{
Returns a one-line tibble containing: reaction sequence up to this point, product molecule (in SMILES format), and the current value of savePoint (TRUE or FALSE).
}
\description{
Predict fragments from specified reaction events on a single molecule.
}
