% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictReactionProducts.R
\name{predict.Hydrolysis}
\alias{predict.Hydrolysis}
\title{Predicts the products of hydrolysis at a phosphate ester linkage.}
\usage{
\method{predict}{Hydrolysis}(rxnSite)
}
\arguments{
\item{rxnSite}{One-row tibble containing: molecule (as SMILES string), reaction site type (string), SDF bondblock row number, and SDF atomblock row numbers for phosphorus, oxygen, sulfur, and/or carbon. See findBonds() for more details.}
}
\value{
A character vector of length two containing SMILES strings, one for each hydrolysis product.
}
\description{
Predicts the products of hydrolysis at a phosphate ester linkage.
}
