% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictReactionProducts.R
\name{predict.Product}
\alias{predict.Product}
\title{Predict reaction products of specified reactions with a single molecule.}
\usage{
\method{predict}{Product}(cmpd, grp, rxnList, rxnSave, savePoint)
}
\arguments{
\item{cmpd}{SMILES string representing a single molecule.}

\item{grp}{Single-element named character vector containing a functional group. Name represents atom indices. Output of smartBonds().}

\item{rxnList}{Character vector of allowed reaction types.}

\item{rxnSave}{Character vector containing saved reactions preceding this one in sequence.}

\item{savePoint}{Single-element logical vector. Allows for different reactions at the same reaction site within recursivePredict().}
}
\value{
Returns a one-line tibble containing: reaction sequence up to this point, product molecule (in SMILES format), and the current value of savePoint (TRUE or FALSE).
}
\description{
Predict reaction products of specified reactions with a single molecule.
}
