% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPCFrags.R
\name{recursive.Frag}
\alias{recursive.Frag}
\title{Recursive prediction of transformation products.}
\usage{
recursive.Frag(SMI, prods, funGroups, rxnList, rxnSave, skipCheck)
}
\arguments{
\item{SMI}{Character vector of length one containing a SMILES string.}

\item{prods}{Empty three-column tibble in which to store results.}

\item{funGroups}{Character vector containing functional groups at which transformation reactions can occur.}

\item{rxnList}{Character vector containing types of reactions to use for prediction.}

\item{rxnSave}{Empty character vector for storing sequential reaction information.}

\item{skipCheck}{Empty character vector for storing SMILES strings of predicted products. Used to improve efficiency by not performing the same prediction multiple times.}
}
\value{
Tibble containing fragment SMILES, list of sequential fragmentation reactions, and saved skipCheck strings.
}
\description{
Recursive prediction of transformation products.
}
