% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursivePredict.R
\name{recursive.Predict}
\alias{recursive.Predict}
\title{Recursive prediction of transformation products.}
\usage{
recursive.Predict(
  SMI,
  prods,
  funGroups,
  rxnList,
  rxnSave,
  savePoint,
  skipCheck
)
}
\arguments{
\item{SMI}{Character vector of length one containing a SMILES string.}

\item{prods}{Empty four-column tibble in which to store results.}

\item{funGroups}{Character vector containing functional groups at which transformation reactions can occur.}

\item{rxnList}{Character vector containing types of reactions to use for prediction.}

\item{rxnSave}{Empty character vector for storing sequential reaction information.}

\item{savePoint}{Single logical value. Value determines whether to perform hydrolysis or ozonolysis on a given alkene group. Recursion allows for both reactions to be performed on a single alkene group.}

\item{skipCheck}{Empty character vector for storing SMILES strings of predicted products. Used to improve efficiency by not performing the same prediction multiple times.}
}
\value{
Tibble containing product SMILES, list of sequential reactions for each product, and saved skipCheck strings.
}
\description{
Recursive prediction of transformation products.
}
