% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartBonds.R
\name{smartBonds}
\alias{smartBonds}
\title{Identify desired functional groups in a molecule (in SMILES format) and return a vector of functional group strings named with atom indices (Structure-Data Format)}
\usage{
smartBonds(smiles, grps)
}
\arguments{
\item{smiles}{A SMILES-formatted molecule string.}

\item{grps}{A predefined character vector of functional groups.}
}
\value{
A named character vector.
}
\description{
Identify desired functional groups in a molecule (in SMILES format) and return a vector of functional group strings named with atom indices (Structure-Data Format)
}
\examples{
x <- "CCC=CCCO[P](OCCC)(=O)OCCCC"
y <- c("C=C", "PR3", "P=S", "ester_PO")
smartBonds(x, y)
}
