% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonds.R
\name{smartBonds}
\alias{smartBonds}
\title{Quickly detects specific functional groups in a molecule}
\usage{
smartBonds(smiles, grps)
}
\arguments{
\item{smiles}{A SMILES-formatted molecule string.}

\item{grps}{A predefined character vector of functional groups.}
}
\value{
A named character vector.
}
\description{
Detects specific functional groups in a molecule (SMILES format) and returns a vector of functional group strings named with atom indices (Structure-Data Format). Faster and more lightweight than findBonds().
}
